Move Ordebook in autopilot::infra
