+++ crates/autopilot/src/infra/mod.rs
@@ -1,2 +1,5 @@
 pub mod blockchain;
+pub mod orderbook;
 pub mod persistence;
+
+pub use orderbook::Orderbook;
