+++ crates/autopilot/src/protocol/mod.rs
@@ -2,6 +2,3 @@
 //! components.
 
 pub mod fee;
-pub mod orderbook;
-
-pub use self::orderbook::Orderbook;
