+++ crates/autopilot/src/run.rs
@@ -13,7 +13,6 @@ use {
         driver_api::Driver,
         event_updater::{EventUpdater, GPv2SettlementContract},
         infra::{self, blockchain},
-        protocol,
         run_loop::RunLoop,
         shadow,
         solvable_orders::SolvableOrdersCache,
@@ -623,7 +622,7 @@ pub async fn run(args: Arguments) {
 async fn shadow_mode(args: Arguments) -> ! {
     let http_factory = HttpClientFactory::new(&args.http_client);
 
-    let orderbook = protocol::Orderbook::new(
+    let orderbook = infra::Orderbook::new(
         http_factory.create(),
         args.shadow.expect("missing shadow mode configuration"),
     );
