+++ crates/autopilot/src/shadow.rs
@@ -16,7 +16,8 @@ use {
             reveal,
             solve::{self},
         },
-        protocol::{self, fee},
+        infra,
+        protocol::fee,
         run_loop::{self, observe},
     },
     ::observe::metrics,
@@ -38,7 +39,7 @@ impl LivenessChecking for Liveness {
 }
 
 pub struct RunLoop {
-    orderbook: protocol::Orderbook,
+    orderbook: infra::Orderbook,
     drivers: Vec<Driver>,
     trusted_tokens: AutoUpdatingTokenList,
     auction: domain::AuctionId,
@@ -50,7 +51,7 @@ pub struct RunLoop {
 
 impl RunLoop {
     pub fn new(
-        orderbook: protocol::Orderbook,
+        orderbook: infra::Orderbook,
         drivers: Vec<Driver>,
         trusted_tokens: AutoUpdatingTokenList,
         score_cap: U256,
